requirements.txt
# Dipendenze principali
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0

# Per ChromaDB - versione compatibile con Python 3.13
chromadb==0.4.24
langchain>=0.0.267
openai>=1.1.1

# Per il database
sqlalchemy>=2.0.21
alembic>=1.12.0
asyncpg>=0.28.0

# Utilities
python-multipart>=0.0.6
httpx>=0.24.1
tenacity>=8.2.3

# Per il client LogService
# Nota: questo sarà installato dalla directory di installazione locale
# Assicurati che la directory PramaIA-LogService/clients/python sia accessibile
